+++ src/query/service/tests/it/storages/fuse/operations/mod.rs
@@ -24,5 +24,6 @@ mod optimize;
 mod purge_drop;
 mod purge_truncate;
 mod read_plan;
+mod replace_into;
 mod table_analyze;
 mod truncate;
