+++ src/query/storages/fuse/Cargo.toml
@@ -56,6 +56,3 @@ tokio-rayon = "2.1.0"
 tracing = "0.1.36"
 typetag = "0.2.3"
 uuid = { version = "1.1.2", features = ["serde", "v4"] }
-
-[dev-dependencies]
-rand = "0.8.5"
